Fixed a bug with spaces in expressions and * in evaluate
The first bug was fixed by checking if the comment field in equ/org
lines is not empty if that's the case then the comment is concatenated
with the operand.

The second bug was fixed by adding a third parameter to the evaluate
function which represents the current value of the location counter
if any of the terms in the expression was found to be * its value is set
to the value of the third parameter.